chore(ci): benchmarking
