+++ .github/workflows/test.yml
@@ -3,8 +3,7 @@ name: Node CI
 on: [push, pull_request]
 
 jobs:
-  build:
-
+  test:
     runs-on: ${{ matrix.os }}
 
     strategy:
@@ -14,13 +13,33 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
+
     - name: Use Node.js ${{ matrix.node-version }}
       uses: actions/setup-node@v2
       with:
         node-version: ${{ matrix.node-version }}
+
     - name: npm install, test
       run: |
-        npm install
-        npm test
+        yarn
+        yarn test
+
+    - name: benchmark
+      run: yarn benchmark | tee output.txt
+
+    - name: Download previous benchmark result
+      uses: actions/cache@v1
+      with:
+        path: ./cache
+        key: ${{ runner.os }}-benchmark
+
+    - name: Store benchmark result
+      uses: benchmark-action/github-action-benchmark@v1
+      with:
+        tool: 'benchmarkjs'
+        output-file-path: output.txt
+        external-data-json-path: ./cache/benchmark-data.json
+        alert-threshold: '110%'
+        fail-on-alert: true
       env:
         CI: true
