+++ package.json
@@ -20,7 +20,8 @@
     "-postbuild": "yarpm run build-docs",
     "build-docs": "jsdoc -c .jsdoc.json",
     "build-types": "jsdoc -t node_modules/tsd-jsdoc/dist -r src -d .",
-    "publish-docs": "git subtree push --prefix docs origin gh-pages"
+    "publish-docs": "git subtree push --prefix docs origin gh-pages",
+    "benchmark": "node test/benchmark.js"
   },
   "husky": {
     "hooks": {
@@ -63,6 +64,7 @@
     "@babel/core": "^7.16.12",
     "@babel/preset-env": "^7.16.11",
     "babel-loader": "^8.0.5",
+    "benchmark": "^2.1.4",
     "coveralls": "^3.0.3",
     "eslint": "^8.8.0",
     "eslint-plugin-jest": "^26.0.0",
